corrigindo dados arquivos

+++ src/main/java/Model/IImportarDados.java
@@ -8,7 +8,7 @@ import javax.swing.JFileChooser;
 
 public interface IImportarDados {
     
-    public ArrayList<Integer> importarDados(File arquivo);
+    public ArrayList<Double> importarDados(File arquivo);
     public boolean verificaArquivo(JFileChooser fileChooser, int result);
